+++ src/is_empty.rs
@@ -3,7 +3,7 @@ use crate::bitset::BitSet;
 impl BitSet {
     /// Returns true if the set contains no elements.
     pub fn is_empty(&self) -> bool {
-        let mut result: u64 = 0;
+        let mut result: usize = 0;
         for bits in self.0 {
             result &= bits;
         }
