+++ src/total_set.rs
@@ -4,6 +4,6 @@ use crate::bitset::BitSet;
 impl BitSet {
     /// Creates a new bitset that contains every u8 value.
     pub fn total_set() -> Self {
-        Self([u64::MAX; BITSET_ARRAY_SIZE])
+        Self([usize::MAX; BITSET_ARRAY_SIZE])
     }
 }
